requirements.txt
fastapi
selenium
uvicorn
webdriver-manager
sqlalchemy
"python-jose[cryptography]"
python-dotenv
python-multipart
"passlib[bcrypt]"
scikit-learn
pandas
prometheus_fastapi_instrumentator
streamlit